Refactor from personal review of #215
