+++ lib/app_registry.js
@@ -51,7 +51,7 @@ import { getStorybookUI, configure } from '@kadira/react-native-storybook';
 
 // import your stories
 configure(() => {
-  require('./stories/bio');
+  require('./stories/artist_component');
 }, module);
 
 const StorybookUI = getStorybookUI({port: 9001, host: 'localhost'});
