+++ lib/stories/artist_component.js
